tree: move ctrl-find cluster to tree-fabrics.c
Move the ctrl-find/match cluster out of tree.c and into tree-fabrics.c
so that tree.c no longer depends on private-fabrics.h.

Promote traddr_is_hostname, libnvmf_default_config, and
libnvme_ctrl_find declarations from private-fabrics.h to private.h
since they are called from tree.c which no longer includes
private-fabrics.h.
