+++ libnvme/src/nvme/private-fabrics.h
@@ -183,23 +183,6 @@ struct candidate_args {
 typedef bool (*ctrl_match_t)(struct libnvme_ctrl *c,
 		struct candidate_args *candidate);
 
-bool libnvme_tree_ctrl_match(struct libnvme_ctrl *c,
-		struct candidate_args *candidate);
-ctrl_match_t libnvmf_candidate_init(struct libnvme_global_ctx *ctx,
-		struct candidate_args *candidate,
-		const struct libnvme_ctrl_params *params);
-
-bool traddr_is_hostname(struct libnvme_global_ctx *ctx,
-		const char *transport, const char *traddr);
-
-void libnvmf_default_config(struct libnvme_fabrics_config *cfg);
-
-ctrl_match_t libnvme_candidate_init(struct libnvme_global_ctx *ctx,
-		struct candidate_args *candidate,
-		const struct libnvme_ctrl_params *params);
-libnvme_ctrl_t libnvme_ctrl_find(libnvme_subsystem_t s,
-		const struct libnvme_ctrl_params *params, libnvme_ctrl_t p);
-
 bool libnvmf_ctrl_match_config(struct libnvme_ctrl *c,
 		struct libnvmf_context *fctx);
 struct libnvme_ctrl *libnvmf_ctrl_find(struct libnvme_subsystem *s,
